Initial TypeScript port of authentication-local
