+++ packages/authentication-local/.npmignore
@@ -1,1 +1,3 @@
 test/
+src/
+tsconfig.json
