+++ packages/authentication-local/package.json
@@ -26,7 +26,11 @@
     "node": ">= 6"
   },
   "scripts": {
-    "test": "mocha --opts ../../mocha.opts"
+    "prepublish": "npm run compile",
+    "lint": "tslint --fix src/**/* test/**/*",
+    "compile": "shx rm -rf lib/ && tsc",
+    "test": "npm run lint && npm run mocha",
+    "mocha": "mocha --opts ../../mocha.ts.opts --recursive test/**.test.ts test/**/*.test.ts"
   },
   "directories": {
     "lib": "lib"
@@ -43,6 +47,14 @@
   "devDependencies": {
     "@feathersjs/authentication": "^2.1.16",
     "@feathersjs/feathers": "^3.3.1",
-    "mocha": "^6.0.2"
+    "@types/bcryptjs": "^2.4.2",
+    "@types/debug": "^4.1.3",
+    "@types/lodash": "^4.14.123",
+    "@types/mocha": "^5.2.6",
+    "@types/node": "^11.12.2",
+    "mocha": "^6.0.2",
+    "ts-node": "^8.0.3",
+    "tslint": "^5.14.0",
+    "typescript": "^3.4.1"
   }
 }
