+++ packages/authentication/.npmignore
@@ -1,1 +1,3 @@
 test/
+src/
+tsconfig.json
